requirements.txt
fastapi
Pillow
numpy
torch
torchvision
uvicorn
opencv-python
matplotlib
python-multipart